Sweet alerts para todo lo de la ventana de unidad funcional

+++ Frontend/static/js/registrar_pagos.js
@@ -10,9 +10,9 @@ document.addEventListener('DOMContentLoaded', () => {
     const tenant = btnFormularioPagos.getAttribute('data-tenant')
 
     btnFormularioPagos.addEventListener('click', async () => {
-        modalOverlay.style.display = 'flex';
         const response = await fetch(`/pagos/${consortiumId}/${unitId}/${tenant}`);
         const pagos = await response.json();
+        modalOverlay.style.display = 'flex';
 
         tablaPagosBody.innerHTML = '';
 
@@ -34,12 +34,50 @@ document.addEventListener('DOMContentLoaded', () => {
 
     btnRegistrarPagos.addEventListener('click', async () => {
         const checkboxes = document.querySelectorAll('.pago-checkbox:checked');
-        for (const checkbox of checkboxes) {
-            const paymentId = checkbox.getAttribute('data-payment-id');
-            await fetch(`/registrar_pago/${paymentId}`, { method: 'DELETE' });
+        const cantidad = checkboxes.length;
+        if (cantidad < 1) {
+            Swal.fire({
+                title: "Error",
+                text: "Selecciona al menos un pago.",
+                icon: "error"
+            });
+        } else {
+            Swal.fire({
+                title: "¿Seguro?",
+                text: `Estás a punto de registrar ${cantidad} pagos`,
+                icon: "question",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Si"
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    let huboError = false;
+                    for (const checkbox of checkboxes) {
+                        const paymentId = checkbox.getAttribute('data-payment-id');
+                        const response = await fetch(`/registrar_pago/${paymentId}`, {method: 'DELETE'});
+                        if (!response.ok) {
+                            huboError = true;
+                            break;
+                        }
+                    }
+                    if (huboError) {
+                        Swal.fire({
+                            title: "Error",
+                            text: "Ocurrió un error al registrar los pagos.",
+                            icon: "error"
+                        });
+                    } else {
+                        modalOverlay.style.display = 'none';
+                        Swal.fire({
+                            title: "Hecho",
+                            text: "Los pagos se han registrado.",
+                            icon: "success"
+                        });
+                    }
+                }
+            });
         }
-        modalOverlay.style.display = 'none';
-        location.reload();
     });
 
     btnCancelarRegistroPago.addEventListener('click', () => {
